Changes related to autoCosting calculations and delAudit.

+++ Bsdl.FreshTrade.Repositories.PreInv/dto/PrcCstInfo.dto.cs
@@ -80,7 +80,7 @@ namespace Bsdl.FreshTrade.Repositories.PreInv.dto
 
     
        [DataMember]              
-        public decimal? Prcpriceisper {get;set;}
+        public int? Prcpriceisper {get;set;}
 
     
        [DataMember]
